fix bugs

+++ src/symlink.rs
@@ -8,6 +8,8 @@ pub fn link<P: AsRef<Path>, U: AsRef<Path>>(original: P, link: U) -> std::io::Re
 
 #[cfg(unix)]
 pub fn remove<P: AsRef<Path>>(symlink_file: P) -> std::io::Result<()> {
-    std::fs::remove_file(symlink_file)?;
+    if std::fs::symlink_metadata(&symlink_file).is_ok() {
+        std::fs::remove_file(symlink_file)?;
+    }
     Ok(())
 }
